@@ -42,7 +42,7 @@
     "alt-container": "^1.0.0",
     "alt-react": "0.0.1",
     "bignumber.js": "^4.0.0",
-    "bitsharesjs": "^1.2.4",
+    "bitsharesjs": "^1.2.7",
     "classnames": "^2.2.1",
     "cookies-js": "^1.2.1",
     "counterpart": "^0.17.1",
@@ -59,14 +59,13 @@
     "lodash": "^3.10.1",
     "lzma": "2.1.6",
     "node-fetch": "^1.3.1",
-    "numeral": "^2.0.6",
+    "numeral": "2.0.4",
     "object-assign": "^4.0.1",
-    "peerjs": "^0.3.14",
     "perfect-scrollbar": "git+https://github.com/bitshares/perfect-scrollbar.git",
-    "react": "^15.4.2",
-    "react-addons-css-transition-group": "^15.4.2",
+    "react": "^15.6.1",
+    "react-addons-css-transition-group": "^15.6.0",
     "react-clipboard.js": "^1.0.1",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.6.1",
     "react-foundation-apps": "git+https://github.com/valzav/react-foundation-apps.git",
     "react-highcharts": "^10.0.0",
     "react-hot-loader": "^3.0.0-beta.6",
@@ -87,8 +86,7 @@
     "babel-core": "^6.21.0",
     "babel-eslint": "^7.1.1",
     "babel-jest": "^18.0.0",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-loader": "^7.1.1",
     "babel-plugin-webpack-alias": "^2.1.2",
     "babel-polyfill": "^6.20.0",
     "babel-preset-es2015": "^6.18.0",
@@ -106,7 +104,7 @@
     "eslint-plugin-react": "^5.2.2",
     "express": "^4.14.0",
     "express-history-api-fallback": "^2.1.0",
-    "extract-text-webpack-plugin": "^2.0.0-beta.4",
+    "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.8.4",
     "git-rev-sync": "^1.8.0",
     "html-loader": "^0.4.4",
@@ -115,19 +113,19 @@
     "json-loader": "^0.5.4",
     "mocha": "^3.2.0",
     "node-libs-browser": "^1.0.0",
-    "node-sass": "^3.4.1",
+    "node-sass": "^4.5.3",
     "postcss-loader": "^0.8.2",
     "remarkable-loader": "^0.2.1",
-    "sass-loader": "^4.1.1",
+    "sass-loader": "^6.0.6",
     "script-loader": "^0.6.1",
     "style-loader": "^0.13.1",
     "svg-inline-loader": "^0.7.1",
     "svgo-loader": "1.1.2",
     "url-loader": "^0.5.7",
-    "webpack": "^2.2.0",
-    "webpack-dev-middleware": "^1.9.0",
+    "webpack": "^3.5.4",
+    "webpack-dev-middleware": "^1.12.0",
     "webpack-hot-middleware": "^2.14.0",
-    "worker-loader": "^0.7.1"
+    "worker-loader": "^0.8.1"
   },
   "jest": {
     "scriptPreprocessor": "<rootDir>/__tests__/jest-preprocessor.js",
