@@ -56,25 +56,24 @@ class App extends React.Component {
             NotificationStore.listen(this._onNotificationChange.bind(this));
             SettingsStore.listen(this._onSettingsChange.bind(this));
 
-            // ChainStore.init().then(() => {
 
-            Promise.all([
-                AccountStore.loadDbData(Apis.instance().chainId)
-            ]).then(() => {
-                window._debug_wss_set = function(set){
-                  Apis.instance().ws_rpc.ws.debug = !Apis.instance().ws_rpc.ws.debug;
-                }
-                AccountStore.tryToSetCurrentAccount();
-                this.setState({loading: false});
+            ChainStore.init().then(() => {
+                this.setState({synced: true});
+                Promise.all([
+                    AccountStore.loadDbData(Apis.instance().chainId)
+                ]).then(() => {
+                    AccountStore.tryToSetCurrentAccount();
+                    this.setState({loading: false});
+                }).catch(error => {
+                    console.log("[App.jsx] ----- ERROR ----->", error);
+                    this.setState({loading: false});
+                });
             }).catch(error => {
-                console.log("[App.jsx] ----- ERROR ----->", error);
-                this.setState({loading: false});
+                console.log("[App.jsx] ----- ChainStore.init error ----->", error);
+                let syncFail = ChainStore.subError && (ChainStore.subError.message === "ChainStore sync error, please check your system clock") ? true : false;
+
+                this.setState({loading: false, synced: false, syncFail});
             });
-            // }).catch(error => {
-            //     console.log("[App.jsx] ----- ChainStore.init error ----->", error);
-            //
-            //     this.setState({loading: false, syncFail});
-            // });
         } catch(e) {
             console.error("e:", e);
         }
@@ -235,6 +234,17 @@ class Root extends React.Component {
     static childContextTypes = {
         router: React.PropTypes.object,
         location: React.PropTypes.object
+    }
+
+    componentDidMount(){
+        //Detect OS for platform specific fixes
+        if(navigator.platform.indexOf('Win') > -1){
+            var main = document.getElementById('content');
+            var windowsClass = 'windows';
+            if(main.className.indexOf('windows') === -1){
+                main.className = main.className + (main.className.length ? ' ' : '') + windowsClass;
+            }
+        }
     }
 
     getChildContext() {
