@@ -1,6 +1,6 @@
 import WalletUnlockActions from "actions/WalletUnlockActions";
 import WalletDb from "stores/WalletDb";
-import {Aes, ChainValidation, TransactionBuilder, TransactionHelper, FetchChain} from "bitsharesjs/es";
+import {Aes, ChainValidation, TransactionBuilder, TransactionHelper, ops, FetchChain, ChainStore} from "bitsharesjs/es";
 
 class ApplicationApi {
 
@@ -278,6 +278,40 @@ class ApplicationApi {
 
             return WalletDb.process_transaction(tr, null, true)
         })
+    }
+
+    createWorker(options, account) {
+
+        console.log("template:", TransactionHelper.template("worker_create"));
+        return new Promise((resolve, reject) => {
+            let tr = new TransactionBuilder();
+            console.log("ops:", ops["worker_initializer"], ops["vesting_balance_worker_initializer"].toObject(undefined));
+            const core = ChainStore.getAsset("1.3.0");
+            if (!core) reject(new Error("Can't find core asset, please try again"));
+            let precision = Math.pow(10, core.get("precision"));
+
+            const owner = ChainStore.getAccount(account).get("id");
+            if (!owner) reject(new Error("Can't find the owner account, please try again"));
+            
+            try {
+                tr.add_type_operation("worker_create", {
+                    fee: {
+                        amount: 0,
+                        asset_id: 0
+                    },
+                    owner,
+                    work_begin_date: options.start,
+                    work_end_date: options.end,
+                    daily_pay: options.pay * precision,
+                    name: options.title,
+                    url: options.url,
+                    initializer: [1, {pay_vesting_period_days: options.vesting}]
+                });
+            } catch(err) {
+                reject(err);
+            }
+            WalletDb.process_transaction(tr, null, true).then(resolve).catch(reject);
+        });
     }
 }
 
