@@ -11,11 +11,11 @@ class LinkToAccountById extends React.Component {
     };
 
     static defaultProps = {
-        subpage: "overview"
+        subpage: "overview",
+        autosubscribe: false
     };
 
     shouldComponentUpdate(nextProps) {
-        // console.log("linkToAccountById:", nextProps.account.toJS());
         if (nextProps.account.get("name") && this.props.account.get("name") && nextProps.account.get("name") === this.props.account.get("name")) {
             return false;
         }
@@ -25,10 +25,7 @@ class LinkToAccountById extends React.Component {
     render() {
         let account_name = this.props.account.get("name");
         if (!account_name) {
-            console.log( "!account_name: ", this.props.account.toJS() );
             return <span>{this.props.account.get("id")}</span>;
-        } else {
-            // console.log( "account_name exists: ", this.props.account.get("id"), this.props.account.get("name") );
         }
 
         return <Link onClick={this.props.onClick ? this.props.onClick : () => {}} to={`/account/${account_name}/${this.props.subpage}/`}>{account_name}</Link>;
