@@ -8,12 +8,14 @@ import BalanceComponent from "components/Utility/BalanceComponent";
 import counterpart from "counterpart";
 import AmountSelector from "components/Utility/AmountSelector";
 import AccountActions from "actions/AccountActions";
-import Modal from "react-foundation-apps/src/modal";
 import ZfApi from "react-foundation-apps/src/utils/foundation-api";
 import { validateAddress, WithdrawAddresses } from "common/blockTradesMethods";
 import AccountStore from "stores/AccountStore";
 import {ChainStore} from "bitsharesjs/es";
-
+import Modal from "react-foundation-apps/src/modal";
+import { checkFeeStatusAsync, checkBalance } from "common/trxHelper";
+import {Asset} from "common/MarketClasses";
+import { debounce } from "lodash";
 
 class WithdrawModalBlocktrades extends React.Component {
 
@@ -46,25 +48,55 @@ class WithdrawModalBlocktrades extends React.Component {
             withdraw_address_first: true,
             empty_withdraw_value: false,
             from_account: ChainStore.getAccount(AccountStore.getState().currentAccount),
-            from_error: null,
-            asset: null,
-            feeAsset: null,
             fee_asset_id: "1.3.0"
         };
 
+
         this._validateAddress(this.state.withdraw_address, props);
+
+        this._checkBalance = this._checkBalance.bind(this);
+        this._updateFee = debounce(this._updateFee.bind(this), 250);
+    }
+
+    componentWillMount() {
+        this._updateFee();
+    }
+
+    componentWillUnmount() {
+        this.unMounted = true;
+    }
+
+    _updateFee(fee_asset_id = this.state.fee_asset_id) {
+        checkFeeStatusAsync({
+            accountID: this.props.account.get("id"),
+            feeID: fee_asset_id,
+            options: ["price_per_kbyte"],
+            data: {
+                type: "memo",
+                content: this.props.output_coin_type + ":" + this.state.withdraw_address + (this.state.memo ? ":" + this.state.memo : "")
+            }
+        })
+        .then(({fee, hasBalance, hasPoolBalance}) => {
+            if (this.unMounted) return;
+
+            this.setState({
+                feeAmount: fee,
+                hasBalance,
+                hasPoolBalance,
+                error: (!hasBalance || !hasPoolBalance)
+            }, this._checkBalance);
+        });
     }
 
     onMemoChanged( e ) {
-        this.setState( {memo: e.target.value} );
+        this.setState( {memo: e.target.value}, this._updateFee);
     }
 
     onWithdrawAmountChange( {amount} ) {
-
         this.setState({
             withdraw_amount: amount,
-            empty_withdraw_value: !parseFloat(amount)
-        });
+            empty_withdraw_value: amount !== undefined && !parseFloat(amount)
+        }, this._checkBalance);
     }
 
     onSelectChanged(index) {
@@ -105,45 +137,53 @@ class WithdrawModalBlocktrades extends React.Component {
             });
     }
 
+    _checkBalance() {
+        const {feeAmount, withdraw_amount} = this.state;
+        const {asset} = this.props;
+
+        const hasBalance = checkBalance(withdraw_amount, asset, feeAmount, this.props.balance);
+        if (hasBalance === null) return;
+        this.setState({balanceError: !hasBalance});
+        return hasBalance;
+    }
+
     onSubmit() {
 
         if ((!this.state.withdraw_address_check_in_progress) && (this.state.withdraw_address && this.state.withdraw_address.length) && (this.state.withdraw_amount !== null)) {
             if (!this.state.withdraw_address_is_valid) {
                 ZfApi.publish(this.getWithdrawModalId(), "open");
             } else if (parseFloat(this.state.withdraw_amount) > 0){
-
-               if (!WithdrawAddresses.has(this.props.output_wallet_type)) {
-
+                if (!WithdrawAddresses.has(this.props.output_wallet_type)) {
                     let withdrawals = [];
                     withdrawals.push(this.state.withdraw_address);
-
                     WithdrawAddresses.set({wallet: this.props.output_wallet_type, addresses: withdrawals});
-                    // localStorage.setItem(`history_address_${this.props.output_wallet_type}`, JSON.stringify(withdrawals));
                 } else {
-
                     let withdrawals = WithdrawAddresses.get(this.props.output_wallet_type);
-                    // let withdrawals = JSON.parse(localStorage.getItem(`history_address_${this.props.output_wallet_type}`));
                     if (withdrawals.indexOf(this.state.withdraw_address) == -1) {
-
                         withdrawals.push(this.state.withdraw_address);
                         WithdrawAddresses.set({wallet: this.props.output_wallet_type, addresses: withdrawals});
-                        // localStorage.setItem(`history_address_${this.props.output_wallet_type}`, JSON.stringify(withdrawals));
                     }
                 }
                 WithdrawAddresses.setLast({wallet: this.props.output_wallet_type, address: this.state.withdraw_address});
-                // lStorage.setItem(`history_address_last_${this.props.output_wallet_type}`, this.state.withdraw_address);
                 let asset = this.props.asset;
-                let precision = utils.get_asset_precision(asset.get("precision"));
-                let amount = String.prototype.replace.call(this.state.withdraw_amount, /,/g, "");
+
+                const {feeAmount} = this.state;
+
+                let amount = parseFloat(String.prototype.replace.call(this.state.withdraw_amount, /,/g, ""));
+                let sendAmount = new Asset({
+                    asset_id: asset.get("id"),
+                    precision: asset.get("precision"),
+                    real: amount
+                });
 
                 AccountActions.transfer(
                     this.props.account.get("id"),
                     this.props.issuer.get("id"),
-                    parseInt(amount * precision, 10),
+                    sendAmount.getAmount(),
                     asset.get("id"),
                     this.props.output_coin_type + ":" + this.state.withdraw_address + (this.state.memo ? ":" + new Buffer(this.state.memo, "utf-8") : ""),
                     null,
-                    this.state.feeAsset ? this.state.feeAsset.get("id") : "1.3.0"
+                    feeAmount ? feeAmount.asset_id : "1.3.0"
                 );
 
                 this.setState({
@@ -165,22 +205,19 @@ class WithdrawModalBlocktrades extends React.Component {
             let withdrawals = [];
             withdrawals.push(this.state.withdraw_address);
             WithdrawAddresses.set({wallet: this.props.output_wallet_type, addresses: withdrawals});
-            // localStorage.setItem(`history_address_${this.props.output_wallet_type}`, JSON.stringify(withdrawals));
-
         } else {
             let withdrawals = WithdrawAddresses.get(this.props.output_wallet_type);
-            // let withdrawals = JSON.parse(localStorage.getItem(`history_address_${this.props.output_wallet_type}`));
             if (withdrawals.indexOf(this.state.withdraw_address) == -1) {
                 withdrawals.push(this.state.withdraw_address);
                 WithdrawAddresses.set({wallet: this.props.output_wallet_type, addresses: withdrawals});
-                // localStorage.setItem(`history_address_${this.props.output_wallet_type}`, JSON.stringify(withdrawals));
             }
         }
         WithdrawAddresses.setLast({wallet: this.props.output_wallet_type, address: this.state.withdraw_address});
-        // localStorage.setItem(`history_address_last_${this.props.output_wallet_type}`, this.state.withdraw_address);
         let asset = this.props.asset;
         let precision = utils.get_asset_precision(asset.get("precision"));
         let amount = String.prototype.replace.call(this.state.withdraw_amount, /,/g, "");
+
+        const {feeAmount} = this.state;
 
         AccountActions.transfer(
             this.props.account.get("id"),
@@ -189,7 +226,7 @@ class WithdrawModalBlocktrades extends React.Component {
             asset.get("id"),
             this.props.output_coin_type + ":" + this.state.withdraw_address + (this.state.memo ? ":" + new Buffer(this.state.memo, "utf-8") : ""),
             null,
-            this.state.feeAsset ? this.state.feeAsset.get("id") : "1.3.0"
+            feeAmount ? feeAmount.asset_id : "1.3.0"
         );
     }
 
@@ -212,21 +249,24 @@ class WithdrawModalBlocktrades extends React.Component {
         return "confirmation";
     }
 
-    onAccountBalance({fee,feeID,total_precision}) {
-
+    onAccountBalance() {
+        const { feeAmount } = this.state;
         if (Object.keys(this.props.account.get("balances").toJS()).includes(this.props.asset.get("id")) ) {
-            let total_minus_fee = this.props.balance.get("balance") / utils.get_asset_precision(this.props.asset.get("precision"));
-            if(this.props.asset.get("id")==feeID){
-                total_minus_fee=utils.limitByPrecision(total_minus_fee-fee*1.095,total_precision); //@#>
-                if(total_minus_fee<0){
-                    total_minus_fee=0;
-                }
+            let total = new Asset({
+                amount: this.props.balance.get("balance"),
+                asset_id: this.props.asset.get("id"),
+                precision: this.props.asset.get("precision")
+            });
+
+            // Subtract the fee if it is using the same asset
+            if(total.asset_id === feeAmount.asset_id) {
+                total.minus(feeAmount);
             }
 
             this.setState({
-                withdraw_amount: total_minus_fee,
+                withdraw_amount: total.getAmount({real: true}),
                 empty_withdraw_value: false
-            });
+            }, this._checkBalance);
         }
     }
 
@@ -234,49 +274,38 @@ class WithdrawModalBlocktrades extends React.Component {
         this.nestedRef = ref;
     }
 
-    _setTotal(asset_id, balance_id, fee, fee_asset_id) {
-
-        const gateFee = parseInt(this.props.gateFee);
-
-        let balanceObject = ChainStore.getObject(balance_id);
-        let transferAsset = ChainStore.getObject(asset_id);
-        if (balanceObject) {
-            let amount = (utils.get_asset_amount(balanceObject.get("balance"), transferAsset) - gateFee - (asset_id === fee_asset_id ? fee : 0)).toString();
-            this.setState({amount});
-        }
-    }
-
     onFeeChanged({asset}) {
-        this.setState({feeAsset: asset, error: null});
+        this.setState({
+            fee_asset_id: asset.get("id")
+        }, this._updateFee);
     }
 
     _getAvailableAssets(state = this.state) {
-        const { from_account, from_error } = state;
-        let asset_types = [];
+        const { from_account } = state;
         let fee_asset_types = [];
-
-        if (!(from_account && from_account.get("balances") && !from_error)) {
-            return {asset_types, fee_asset_types};
+        if (!(from_account && from_account.get("balances"))) {
+            return {fee_asset_types};
         }
         let account_balances = state.from_account.get("balances").toJS();
-
+        fee_asset_types = Object.keys(account_balances).sort(utils.sortID);
         for (let key in account_balances) {
             let asset = ChainStore.getObject(key);
-            let balanceObject = account_balances[key]?ChainStore.getObject(account_balances[key]):null;
-
-            if (balanceObject && balanceObject.get("balance") > 0) {
-                if(fee_asset_types.indexOf(key)==-1){
-                    asset_types.push(key);
+            let balanceObject = ChainStore.getObject(account_balances[key]);
+            if (balanceObject && balanceObject.get("balance") === 0) {
+                if (fee_asset_types.indexOf(key) !== -1) {
+                    fee_asset_types.splice(fee_asset_types.indexOf(key), 1);
                 }
             }
 
-            if(asset&&utils.isValidPrice(asset.getIn(["options", "core_exchange_rate"]))&&parseInt(asset.getIn(["dynamic", "fee_pool"]), 10)>(this._feeBTS||2200)){
-                fee_asset_types.push(key);
+            if (asset) {
+                // Remove any assets that do not have valid core exchange rates
+                if (asset.get("id") !== "1.3.0" && !utils.isValidPrice(asset.getIn(["options", "core_exchange_rate"]))) {
+                    fee_asset_types.splice(fee_asset_types.indexOf(key), 1);
+                }
             }
-
         }
-        
-        return {asset_types, fee_asset_types};
+
+        return {fee_asset_types};
     }
 
     render() {
@@ -285,11 +314,7 @@ class WithdrawModalBlocktrades extends React.Component {
         let storedAddress = WithdrawAddresses.get(this.props.output_wallet_type);
         let balance = null;
 
-        console.log('@>',this.props)
-
-        // console.log( "account: ", this.props.account.toJS() );
         let account_balances = this.props.account.get("balances").toJS();
-        // console.log( "balances: ", account_balances );
         let asset_types = Object.keys(account_balances);
 
         let withdrawModalId = this.getWithdrawModalId();
@@ -301,7 +326,7 @@ class WithdrawModalBlocktrades extends React.Component {
             options =
                 <div className={!storedAddress.length ? "blocktrades-disabled-options" : "blocktrades-options"}>
                     {storedAddress.map(function(name, index){
-                    return <a key={index} onClick={this.onSelectChanged.bind(this, index)}>{name}</a>;
+                        return <a key={index} onClick={this.onSelectChanged.bind(this, index)}>{name}</a>;
                     }, this)}
                 </div>;
         }
@@ -310,8 +335,8 @@ class WithdrawModalBlocktrades extends React.Component {
         {
             if (!this.state.withdraw_address_is_valid) {
 
-            invalid_address_message = <div className="has-error" style={{paddingTop: 10}}><Translate content="gateway.valid_address" coin_type={this.props.output_coin_type} /></div>;
-            confirmation =
+                invalid_address_message = <div className="has-error" style={{paddingTop: 10}}><Translate content="gateway.valid_address" coin_type={this.props.output_coin_type} /></div>;
+                confirmation =
                 <Modal id={withdrawModalId} overlay={true}>
                     <Trigger close={withdrawModalId}>
                         <a href="#" className="close-button">&times;</a>
@@ -346,73 +371,38 @@ class WithdrawModalBlocktrades extends React.Component {
                 </div>;
         }
 
-
         // Estimate fee VARIABLES
-        const { from_account, from_error, fee_asset_id } = this.state;
-        let feeAsset = this.state.feeAsset;
-        let asset = this.state.asset;
-        let fee = this._feeBTS = utils.estimateFee("transfer", null, ChainStore.getObject("2.0.0"));
-        let { fee_asset_types } = this._getAvailableAssets();
-        let balance_fee = null;
-        let feeID = feeAsset ? feeAsset.get("id") : "1.3.0";
-        let core = ChainStore.getObject("1.3.0");
-
-        // Estimate fee      
-
-        if (from_account && from_account.get("balances") && !from_error) {
-
-            let account_balances = from_account.get("balances").toJS();
-
-            // Finish fee estimation            
-            if (feeAsset && feeAsset.get("id") !== "1.3.0" && core) {
-
-                let price = utils.convertPrice(core, feeAsset.getIn(["options", "core_exchange_rate"]).toJS(), null, feeAsset.get("id"));
-                fee = utils.convertValue(price, fee, core, feeAsset);
-
-                if (parseInt(fee, 10) !== fee) {
-                    fee += 1; // Add 1 to round up;
-                }
-            }
-            if (core) {
-                fee = utils.limitByPrecision(utils.get_asset_amount(fee, feeAsset || core), feeAsset ? feeAsset.get("precision") : core.get("precision"));
-            }
-
-            if (asset_types.length === 1) asset = ChainStore.getAsset(asset_types[0]);
-            if (asset_types.length > 0) {
-                let current_asset_id = asset ? asset.get("id") : asset_types[0];                
-                balance_fee = (<span style={{borderBottom: "#A09F9F 1px dotted", cursor: "pointer"}} onClick={this._setTotal.bind(this, current_asset_id, account_balances[current_asset_id], fee, feeID)}><Translate component="span" content="transfer.available"/>: <BalanceComponent balance={account_balances[current_asset_id]}/></span>);
-            } else {
-                balance_fee = "No funds";
-            }
-        } else {
-            fee_asset_types = ["1.3.0"];
-            if (core) {
-                fee = utils.limitByPrecision(utils.get_asset_amount(fee, feeAsset || core), feeAsset ? feeAsset.get("precision") : core.get("precision"));
-            }
-        }
-
-        let total_precision = feeAsset ? feeAsset.get("precision") : core.get("precision");
+        let { fee_asset_types } = this._getAvailableAssets();console.log('@>fee_asset_types11',fee_asset_types)
 
         if (asset_types.length > 0) {
             let current_asset_id = this.props.asset.get("id");
-            if( current_asset_id )
-                balance = (<span><Translate component="span" content="transfer.available"/>: <span className="set-cursor" onClick={()=>{this.onAccountBalance({fee,feeID,total_precision})}}><BalanceComponent balance={account_balances[current_asset_id]}/></span></span>);
+            if( current_asset_id ){
+                let current = account_balances[current_asset_id];
+                balance = (
+                    <span style={{borderBottom: "#A09F9F 1px dotted", cursor: "pointer"}}>
+                        <Translate component="span" content="transfer.available"/>&nbsp;:&nbsp;
+                        <span className="set-cursor" onClick={this.onAccountBalance.bind(this)}>
+                            {current ? <BalanceComponent balance={account_balances[current_asset_id]}/> : 0}
+                        </span>
+                    </span>
+                );}
             else
                 balance = "No funds";
         } else {
             balance = "No funds";
         }
 
-        let fee_asset_choosen="1.3.0";
-        if(this.state.feeAsset){
-            fee_asset_choosen=this.state.feeAsset.get("id");
-        }
+    console.log('@>',this.state.feeAmount&&this.state.feeAmount.getAmount({real: true}));
+    this.state.feeAmount&&console.log('@>this.state.feeAmount.asset_id',this.state.feeAmount.asset_id)
+
 
         return (<form className="grid-block vertical full-width-content">
             <div className="grid-container">
                 <div className="content-block">
                     <h3><Translate content="gateway.withdraw_coin" coin={this.props.output_coin_name} symbol={this.props.output_coin_symbol} /></h3>
                 </div>
+
+                {/* Withdraw amount */}
                 <div className="content-block">
                     <AmountSelector label="modal.withdraw.amount"
                         amount={this.state.withdraw_amount}
@@ -421,28 +411,35 @@ class WithdrawModalBlocktrades extends React.Component {
                         placeholder="0.0"
                         onChange={this.onWithdrawAmountChange.bind(this)}
                         display_balance={balance}
-                    /> 
-                    {this.state.empty_withdraw_value ? <p className="error_grey no-margin" style={{paddingTop: 4}}><Translate content="transfer.errors.valid" /></p>:null}
+                    />
+                    {this.state.empty_withdraw_value ? <p className="has-error no-margin" style={{paddingTop: 10}}><Translate content="transfer.errors.valid" /></p>:null}
+                    {this.state.balanceError ? <p className="has-error no-margin" style={{paddingTop: 10}}><Translate content="transfer.errors.insufficient" /></p>:null}
                 </div>
-                <div className="content-block gate_fee">                    
-                    <AmountSelector
+
+                {/* Fee selection */}
+                {this.state.feeAmount?<div className="content-block gate_fee">
+                    <AmountSelector 
                         refCallback={this.setNestedRef.bind(this)}
                         label="transfer.fee"
                         disabled={true}
-                        amount={fee}
+                        amount={this.state.feeAmount.getAmount({real: true})}
                         onChange={this.onFeeChanged.bind(this)}
-                        asset={fee_asset_choosen}
+                        asset={this.state.feeAmount.asset_id}
                         assets={fee_asset_types}
-                        tabIndex={tabIndex++}
-                    />
-                    {this.props.gateFee?
-                        (<div className="amount-selector right-selector">
-                            <label className="left-label">Gate fee</label>
-                            <div className="inline-label input-wrapper">
-                                <p className="right-selector-input" >{this.props.gateFee}</p>
-                            </div>
-                        </div>):null}
-                </div>
+                        tabIndex={tabIndex++} 
+                    /> 
+                    {this.props.gateFee? (
+                    <div className="amount-selector right-selector">
+                        <label className="left-label">Gate fee</label>
+                        <div className="inline-label input-wrapper">
+                            <p className="right-selector-input">{this.props.gateFee}</p>
+                        </div>
+                    </div>):null}
+                    {!this.state.hasBalance ? <p className="has-error no-margin" style={{paddingTop: 10}}><Translate content="transfer.errors.noFeeBalance" /></p> : null}
+                    {!this.state.hasPoolBalance ? <p className="has-error no-margin" style={{paddingTop: 10}}><Translate content="transfer.errors.noPoolBalance" /></p> : null}
+                </div>: null}
+
+
 
                 <div className="content-block">
                     <label className="left-label">
@@ -459,13 +456,19 @@ class WithdrawModalBlocktrades extends React.Component {
                     </div>
                     {invalid_address_message}
                 </div>
-                {withdraw_memo}
-                <div className="content-block">
 
-                    <a onClick={this.onSubmit.bind(this)} className="button white_color_a"><Translate content="modal.withdraw.submit" /></a>
+                {/* Memo input */}
+                {withdraw_memo}
+
+                {/* Withdraw/Cancel buttons */}
+                <div className="button-group">
+
+                    <div onClick={this.onSubmit.bind(this)} className={"button" + (this.state.error || this.state.balanceError ? (" disabled") : "")}>
+                        <Translate content="modal.withdraw.submit" />
+                    </div>
 
                     <Trigger close={this.props.modal_id}>
-                        <a href className="secondary button"><Translate content="account.perm.cancel" /></a>
+                        <div className="button"><Translate content="account.perm.cancel" /></div>
                     </Trigger>
                 </div>
                 {confirmation}
