@@ -1,7 +1,6 @@
 import React from "react";
 import ZfApi from "react-foundation-apps/src/utils/foundation-api";
-import Modal from "react-foundation-apps/src/modal";
-import Trigger from "react-foundation-apps/src/trigger";
+import BaseModal from "../Modal/BaseModal";
 import utils from "common/utils";
 import Translate from "react-translate-component";
 
@@ -17,26 +16,25 @@ export default class ConfirmModal extends React.Component {
         e.preventDefault();
 
         ZfApi.publish(modalId, "close");
-        this.props.onForce(value);
+        if (value) this.props.onForce();
     }
 
     render() {
-        let {type, diff} = this.props;
+        let {type, diff, hasOrders} = this.props;
 
         return (
-            <Modal id={"modal_confirm_" + type} overlay={true}>
-                <Trigger close={"modal_confirm_" + type}>
-                    <a href="#" className="close-button">&times;</a>
-                </Trigger>
+            <BaseModal id={"modal_confirm_" + type} overlay={true}>
                 <Translate component="h3" content="transaction.confirm" />
                 <div className="grid-block vertical">
-                    <Translate content={"exchange.confirm_" + type} diff={utils.format_number(diff, 2)} />
+                    {!hasOrders ?
+                        <Translate content={"exchange.confirm_no_orders_" + type} />
+                    : <Translate content={"exchange.confirm_" + type} diff={utils.format_number(diff, 2)} />}
                     <div className="button-group" style={{paddingTop: "2rem"}}>
                         <input onClick={this._onForce.bind(this, true)} className="button success" type="submit" value="Yes" />
                         <input onClick={this._onForce.bind(this, false)} className="button info" type="submit" value="No" />
                     </div>
                 </div>
-            </Modal>
+            </BaseModal>
         );
     }
 }
