@@ -40,12 +40,11 @@ class OrderBookRowVertical extends React.Component {
 }
 
 class OrderBookRowHorizontal extends React.Component {
-    shouldComponentUpdate(nextProps) {
-        if (nextProps.order.market_base !== this.props.order.market_base) return false;
+    shouldComponentUpdate(np) {
         return (
-            nextProps.order.ne(this.props.order) ||
-            nextProps.position !== this.props.position ||
-            nextProps.index !== this.props.index
+            np.order.ne(this.props.order) ||
+            np.position !== this.props.position ||
+            np.index !== this.props.index
         );
     }
 
@@ -127,7 +126,7 @@ class OrderBook extends React.Component {
         }
 
         // Change of market or direction
-        if (nextProps.base !== this.props.base || nextProps.quote !== this.props.quote) {
+        if (nextProps.base.get("id") !== this.props.base.get("id") || nextProps.quote.get("id") !== this.props.quote.get("id")) {
             this.setState({
                 scrollToBottom: true
             });
@@ -141,6 +140,17 @@ class OrderBook extends React.Component {
             if (this.refs.bidTransition) {
                 this.refs.bidTransition.resetAnimation();
             }
+
+            if (this.refs.vert_bids) this.refs.vert_bids.scrollTop = 0;
+        }
+
+        if (
+          !utils.are_equal_shallow(nextProps.combinedAsks, this.props.combinedAsks) ||
+          !utils.are_equal_shallow(nextProps.combinedBids, this.props.combinedBids)
+        ) {
+            this.setState({}, () => {
+                this.psUpdate();
+            });
         }
     }
 
@@ -259,7 +269,6 @@ class OrderBook extends React.Component {
         let {showAllAsks, showAllBids, rowCount} = this.state;
 
         let bidRows = null, askRows = null;
-
         if(base && quote) {
             bidRows = combinedBids
             .filter(a => {
