@@ -187,7 +187,7 @@ class Assets extends React.Component {
         return (
             <div className="grid-block vertical">
                 <div className="grid-block page-layout">
-                    <div className="grid-block small-12 medium-10 medium-offset-1 main-content vertical">
+                    <div className="grid-block main-content small-12 medium-10 medium-offset-1 main-content vertical">
                         <div className="generic-bordered-box">
                             <Tabs
                                 tabsClass="no-padding bordered-header"
@@ -197,57 +197,69 @@ class Assets extends React.Component {
                             >
                                 <Tab title="explorer.assets.market">
                                     <div className="grid-block shrink">
-                                        <input style={{maxWidth: "500px"}} placeholder={placeholder} type="text" value={this.state.filterMPA} onChange={this._onFilter.bind(this, "filterMPA")}></input>
+                                        <div className="grid-content">
+                                            <input style={{maxWidth: "500px"}} placeholder={placeholder} type="text" value={this.state.filterMPA} onChange={this._onFilter.bind(this, "filterMPA")}></input>
+                                        </div>
                                     </div>
                                     <div className="grid-block" style={{paddingBottom: 20}}>
-                                        <table className="table">
-                                            <thead>
-                                            <tr>
-                                                <th><Translate component="span" content="explorer.assets.symbol" /></th>
-                                                <th><Translate component="span" content="explorer.assets.issuer" /></th>
-                                                <th><Translate component="span" content="markets.supply" /></th>
-                                                <th></th>
-                                            </tr>
-                                            </thead>
+                                        <div className="grid-content">
+                                            <table className="table">
+                                                <thead>
+                                                <tr>
+                                                    <th><Translate component="span" content="explorer.assets.symbol" /></th>
+                                                    <th><Translate component="span" content="explorer.assets.issuer" /></th>
+                                                    <th><Translate component="span" content="markets.supply" /></th>
+                                                    <th></th>
+                                                </tr>
+                                                </thead>
                                                 <tbody>
-                                                    {mia}
+                                                {mia}
                                                 </tbody>
-                                        </table>
+                                            </table>
+                                        </div>
                                     </div>
                                 </Tab>
 
                                 <Tab title="explorer.assets.user">
                                     <div className="grid-block shrink">
-                                        <input style={{maxWidth: "500px"}} placeholder={placeholder} type="text" value={this.state.filterUIA} onChange={this._onFilter.bind(this, "filterUIA")}></input>
+                                        <div className="grid-content">
+                                            <input style={{maxWidth: "500px"}} placeholder={placeholder} type="text" value={this.state.filterUIA} onChange={this._onFilter.bind(this, "filterUIA")}></input>
+                                        </div>
                                     </div>
                                     <div className="grid-block" style={{paddingBottom: 20}}>
-                                        <table className="table">
-                                            <thead>
-                                            <tr>
-                                                <th><Translate component="span" content="explorer.assets.symbol" /></th>
-                                                <th><Translate component="span" content="explorer.assets.issuer" /></th>
-                                                <th><Translate component="span" content="markets.supply" /></th>
-                                                <th></th>
-                                            </tr>
-                                            </thead>
+                                        <div className="grid-content">
+                                            <table className="table">
+                                                <thead>
+                                                <tr>
+                                                    <th><Translate component="span" content="explorer.assets.symbol" /></th>
+                                                    <th><Translate component="span" content="explorer.assets.issuer" /></th>
+                                                    <th><Translate component="span" content="markets.supply" /></th>
+                                                    <th></th>
+                                                </tr>
+                                                </thead>
 
-                                            <tbody>
+                                                <tbody>
                                                 {uia}
-                                            </tbody>
-                                        </table>
+                                                </tbody>
+                                            </table>
+                                        </div>
                                     </div>
                                 </Tab>
 
                                 <Tab title="explorer.assets.prediction">
                                     <div className="grid-block shrink">
-                                        <input style={{maxWidth: "500px"}} placeholder={counterpart.translate("markets.search").toUpperCase()} type="text" value={this.state.filterPM} onChange={this._onFilter.bind(this, "filterPM")}></input>
+                                        <div className="grid-content">
+                                            <input style={{maxWidth: "500px"}} placeholder={counterpart.translate("markets.search").toUpperCase()} type="text" value={this.state.filterPM} onChange={this._onFilter.bind(this, "filterPM")}></input>
+                                        </div>
                                     </div>
                                     <div className="grid-block" style={{paddingBottom: 20}}>
-                                        <table className="table">
-                                            <tbody>
+                                        <div className="grid-content">
+                                            <table className="table">
+                                                <tbody>
                                                 {pm}
-                                            </tbody>
-                                        </table>
+                                                </tbody>
+                                            </table>
+                                        </div>
                                     </div>
                                 </Tab>
                             </Tabs>
