@@ -237,7 +237,7 @@ class Header extends React.Component {
             </div>
         ) : null;
 
-        let tradeLink = this.props.lastMarket && active.indexOf("market/") === -1 ?
+        let tradeLink = this.props.lastMarket ?
             <a className={cnames({active: active.indexOf("market/") !== -1})} onClick={this._onNavigate.bind(this, `/market/${this.props.lastMarket}`)}><Translate component="span" content="header.exchange" /></a>:
             <a className={cnames({active: active.indexOf("market/") !== -1})} onClick={this._onNavigate.bind(this, "/market/OBITS_BTS")}><Translate component="span" content="header.exchange" /></a>;
 
