@@ -1,8 +1,8 @@
 import React from "react";
 import ZfApi from "react-foundation-apps/src/utils/foundation-api";
-import Modal from "react-foundation-apps/src/modal";
 import Trigger from "react-foundation-apps/src/trigger";
 import Translate from "react-translate-component";
+import BaseModal from "./BaseModal";
 
 export default class BrowserSupportModal extends React.Component {
 
@@ -17,10 +17,7 @@ export default class BrowserSupportModal extends React.Component {
 
     render() {
         return (
-            <Modal id="browser_modal" overlay={true} ref="browser_modal">
-                <Trigger close="browser_modal">
-                    <a href="#" className="close-button">&times;</a>
-                </Trigger>
+            <BaseModal id="browser_modal" overlay={true} ref="browser_modal">
                 <div className="grid-block vertical no-overflow">
                     <Translate component="h3" content="init_error.browser"/>
                     <Translate component="p" content="init_error.browser_text"/>
@@ -33,9 +30,8 @@ export default class BrowserSupportModal extends React.Component {
                             <div className="button"><Translate content="init_error.understand" /></div>
                         </Trigger>
                     </div>
-
                 </div>
-            </Modal>
+            </BaseModal>
         );
     }
 }
