@@ -10,7 +10,7 @@ import WalletSettings from "./WalletSettings";
 import PasswordSettings from "./PasswordSettings";
 import RestoreSettings from "./RestoreSettings";
 import BackupSettings from "./BackupSettings";
-
+import AccessSettings from "./AccessSettings";
 
 class Settings extends React.Component {
 
@@ -73,8 +73,8 @@ class Settings extends React.Component {
         return menuEntries;
     }
 
-    triggerModal(e) {
-        this.refs.ws_modal.show(e);
+    triggerModal(e, ...args) {
+        this.refs.ws_modal.show(e, ...args);
     }
 
     _onChangeSetting(setting, e) {
@@ -218,6 +218,12 @@ class Settings extends React.Component {
             entries = <RestoreSettings passwordLogin={this.props.settings.get("passwordLogin")} />;
             break;
 
+        case "access":
+            entries = <AccessSettings currentNode={settings.get("apiServer")} faucet={settings.get("faucet_address")} nodes={defaults.apiServer} onChange={this._onChangeSetting.bind(this)} apiLatencies={this.props.apiLatencies} triggerModal={this.triggerModal.bind(this)} />;
+            break;
+        case "faucet_address":
+            entries = <input type="text" defaultValue={settings.get("faucet_address")} onChange={this._onChangeSetting.bind(this, "faucet_address")}/>
+            break;
         default:
             entries = settingEntries[activeEntry].map(setting => {
                 return (
@@ -228,9 +234,7 @@ class Settings extends React.Component {
                         defaults={defaults[setting]}
                         onChange={this._onChangeSetting.bind(this)}
                         locales={this.props.localesObject}
-                        triggerModal={this.triggerModal.bind(this)}
                         {...this.state}
-                        apiLatencies={this.props.apiLatencies}
                     />);
             });
             break;
@@ -239,8 +243,6 @@ class Settings extends React.Component {
         return (
             <div className="grid-block page-layout">
                 <div className="grid-block main-content wrap" style={{marginTop: "1rem"}}>
-
-
                     {!traderMode ? null : <div className="grid-content large-offset-2 shrink" >
                         <Translate className="bottom-border" component="h4" content="header.settings" />
                         <ul className="settings-menu">
