@@ -13,6 +13,9 @@ import { RecentTransactions } from "../Account/RecentTransactions";
 import Immutable from "immutable";
 import {ChainStore} from "bitsharesjs/es";
 import {connect} from "alt-react";
+import { checkFeeStatusAsync, checkBalance } from "common/trxHelper";
+import { debounce } from "lodash";
+import { Asset } from "common/MarketClasses";
 
 class Transfer extends React.Component {
 
@@ -38,6 +41,10 @@ class Transfer extends React.Component {
         let currentAccount = AccountStore.getState().currentAccount;
         if (!this.state.from_name) this.state.from_name = currentAccount;
         this.onTrxIncluded = this.onTrxIncluded.bind(this);
+
+        this._updateFee = debounce(this._updateFee.bind(this), 250);
+        this._checkFeeStatus = this._checkFeeStatus.bind(this);
+        this._checkBalance = this._checkBalance.bind(this);
     }
 
     static getInitialState() {
@@ -54,29 +61,33 @@ class Transfer extends React.Component {
             propose: false,
             propose_account: "",
             feeAsset: null,
-            fee_asset_id: "1.3.0"
+            fee_asset_id: "1.3.0",
+            feeAmount: new Asset({amount: 0}),
+            feeStatus: {}
         };
+
     };
 
     componentWillMount() {
         this.nestedRef = null;
+        this._updateFee();
     }
 
     shouldComponentUpdate(np, ns) {
         let { asset_types: current_types } = this._getAvailableAssets();
-        let { asset_types } = this._getAvailableAssets(ns);
+        let { asset_types: next_asset_types } = this._getAvailableAssets(ns);
 
-        if (asset_types.length === 1) {
-            let asset = ChainStore.getAsset(asset_types[0]);
+        if (next_asset_types.length === 1) {
+            let asset = ChainStore.getAsset(next_asset_types[0]);
             if (current_types.length !== 1) {
                 this.onAmountChanged({amount: ns.amount, asset});
             }
 
-            if (asset_types[0] !== this.state.fee_asset_id) {
-                if (asset && this.state.fee_asset_id !== asset_types[0]) {
+            if (next_asset_types[0] !== this.state.fee_asset_id) {
+                if (asset && this.state.fee_asset_id !== next_asset_types[0]) {
                     this.setState({
                         feeAsset: asset,
-                        fee_asset_id: asset_types[0]
+                        fee_asset_id: next_asset_types[0]
                     });
                 }
             }
@@ -88,9 +99,83 @@ class Transfer extends React.Component {
         if (np.currentAccount !== this.state.from_name && np.currentAccount !== this.props.currentAccount) {
             this.setState({
                 from_name: np.currentAccount,
-                from_account: ChainStore.getAccount(np.currentAccount)
-            });
+                from_account: ChainStore.getAccount(np.currentAccount),
+                feeStatus: {},
+                fee_asset_id: "1.3.0",
+                feeAmount: new Asset({amount: 0})
+            }, () => {this._updateFee(); this._checkFeeStatus(ChainStore.getAccount(np.currentAccount));});
         }
+    }
+
+    _checkBalance() {
+        const {feeAmount, amount, from_account, asset} = this.state;
+        if (!asset) return;
+        const balanceID = from_account.getIn(["balances", asset.get("id")]);
+        const feeBalanceID = from_account.getIn(["balances", feeAmount.asset_id]);
+        if (!asset || ! from_account) return;
+        if (!balanceID) return this.setState({balanceError: true});
+        let balanceObject = ChainStore.getObject(balanceID);
+        let feeBalanceObject = feeBalanceID ? ChainStore.getObject(feeBalanceID) : null;
+        if (!feeBalanceObject || feeBalanceObject.get("balance") === 0) {
+            this._updateFee("1.3.0");
+        }
+        const hasBalance = checkBalance(amount, asset, feeAmount, balanceObject);
+        if (hasBalance === null) return;
+        this.setState({balanceError: !hasBalance});
+    }
+
+    _checkFeeStatus(account = this.state.from_account) {
+        if (!account) return;
+
+        const assets = Object.keys(this.state.from_account.get("balances").toJS()).sort(utils.sortID);
+        let feeStatus = {};
+        let p = [];
+        assets.forEach(a => {
+            p.push(checkFeeStatusAsync({
+                accountID: account.get("id"),
+                feeID: a,
+                options: ["price_per_kbyte"],
+                data: {
+                    type: "memo",
+                    content: this.state.memo
+                }
+            }));
+        });
+        Promise.all(p).then(status => {
+            assets.forEach((a, idx) => {
+                feeStatus[a] = status[idx];
+            });
+            if (!utils.are_equal_shallow(this.state.feeStatus, feeStatus)) {
+                this.setState({
+                    feeStatus
+                });
+            }
+            this._checkBalance();
+        }).catch(err => {
+            console.error(err);
+        });
+    }
+
+    _updateFee(fee_asset_id = this.state.fee_asset_id) {
+        if (!this.state.from_account) return null;
+        checkFeeStatusAsync({
+            accountID: this.state.from_account.get("id"),
+            feeID: fee_asset_id,
+            options: ["price_per_kbyte"],
+            data: {
+                type: "memo",
+                content: this.state.memo
+            }
+        })
+        .then(({fee, hasBalance, hasPoolBalance}) => {
+            this.setState({
+                feeAmount: fee,
+                fee_asset_id: fee.asset_id,
+                hasBalance,
+                hasPoolBalance,
+                error: (!hasBalance || !hasPoolBalance)
+            }, this._checkFeeStatus);
+        });
     }
 
     fromChanged(from_name) {
@@ -103,7 +188,7 @@ class Transfer extends React.Component {
     }
 
     onFromAccountChanged(from_account) {
-        this.setState({from_account, error: null});
+        this.setState({from_account, error: null}, this._updateFee);
     }
 
     onToAccountChanged(to_account) {
@@ -114,15 +199,15 @@ class Transfer extends React.Component {
         if (!asset) {
             return;
         }
-        this.setState({amount, asset, asset_id: asset.get("id"), error: null});
+        this.setState({amount, asset, asset_id: asset.get("id"), error: null}, this._checkBalance);
     }
 
     onFeeChanged({asset}) {
-        this.setState({feeAsset: asset, error: null});
+        this.setState({feeAsset: asset, fee_asset_id: asset.get("id"), error: null}, this._updateFee);
     }
 
     onMemoChanged(e) {
-        this.setState({memo: e.target.value});
+        this.setState({memo: e.target.value}, this._updateFee);
     }
 
     onTrxIncluded(confirm_store_state) {
@@ -148,14 +233,13 @@ class Transfer extends React.Component {
     onSubmit(e) {
         e.preventDefault();
         this.setState({error: null});
-        let asset = this.state.asset;
-        let precision = utils.get_asset_precision(asset.get("precision"));
-        let amount = this.state.amount.replace( /,/g, "" );
+        const {asset, amount} = this.state;
+        const sendAmount = new Asset({real: amount, asset_id: asset.get("id"), precision: asset.get("precision")});
 
         AccountActions.transfer(
             this.state.from_account.get("id"),
             this.state.to_account.get("id"),
-            parseInt(amount * precision, 10),
+            sendAmount.getAmount(),
             asset.get("id"),
             this.state.memo ? new Buffer(this.state.memo, "utf-8") : this.state.memo,
             this.state.propose ? this.state.propose_account : null,
@@ -174,41 +258,53 @@ class Transfer extends React.Component {
         this.nestedRef = ref;
     }
 
-    _setTotal(asset_id, balance_id, fee, fee_asset_id) {
+    _setTotal(asset_id, balance_id) {
+        const {feeAmount} = this.state;
         let balanceObject = ChainStore.getObject(balance_id);
         let transferAsset = ChainStore.getObject(asset_id);
+
+        let balance = new Asset({amount: balanceObject.get("balance"), asset_id: transferAsset.get("id"), precision: transferAsset.get("precision")});
+
         if (balanceObject) {
-            let amount = (utils.get_asset_amount(balanceObject.get("balance"), transferAsset) - (asset_id === fee_asset_id ? fee : 0)).toString();
-            this.setState({amount});
+            if (feeAmount.asset_id === balance.asset_id) {
+                balance.minus(feeAmount);
+            }
+            this.setState({amount: balance.getAmount({real: true})}, this._checkBalance);
         }
     }
 
     _getAvailableAssets(state = this.state) {
-        const { from_account, from_error } = state;
-        let asset_types = [];
-        let fee_asset_types = [];
+        const { feeStatus } = this.state;
+        function hasFeePoolBalance(id) {
+            return feeStatus[id] && feeStatus[id].hasPoolBalance;
+        }
 
+        function hasBalance(id) {
+            return feeStatus[id] && feeStatus[id].hasBalance;
+        }
+
+        const { from_account, from_error } = state;
+        let asset_types = [], fee_asset_types = [];
         if (!(from_account && from_account.get("balances") && !from_error)) {
             return {asset_types, fee_asset_types};
         }
         let account_balances = state.from_account.get("balances").toJS();
-
+        asset_types = Object.keys(account_balances).sort(utils.sortID);
+        fee_asset_types = Object.keys(account_balances).sort(utils.sortID);
         for (let key in account_balances) {
-            let asset = ChainStore.getObject(key);
-            let balanceObject = account_balances[key]?ChainStore.getObject(account_balances[key]):null;
-
-            if (balanceObject && balanceObject.get("balance") > 0) {
-                if(fee_asset_types.indexOf(key)==-1){
-                    asset_types.push(key);
+            let balanceObject = ChainStore.getObject(account_balances[key]);
+            if (balanceObject && balanceObject.get("balance") === 0) {
+                asset_types.splice(asset_types.indexOf(key), 1);
+                if (fee_asset_types.indexOf(key) !== -1) {
+                    fee_asset_types.splice(fee_asset_types.indexOf(key), 1);
                 }
             }
-
-            if(asset&&utils.isValidPrice(asset.getIn(["options", "core_exchange_rate"]))&&parseInt(asset.getIn(["dynamic", "fee_pool"]), 10)>(this._feeBTS||2200)){
-                fee_asset_types.push(key);
-            }
-
         }
-        
+
+        fee_asset_types = fee_asset_types.filter(a => {
+            return hasFeePoolBalance(a) && hasBalance(a);
+        });
+
         return {asset_types, fee_asset_types};
     }
 
@@ -224,9 +320,8 @@ class Transfer extends React.Component {
 
     render() {
         let from_error = null;
-
-        let {propose, from_account, to_account, asset, asset_id, propose_account,
-            amount, error, to_name, from_name, memo, feeAsset, fee_asset_id} = this.state;
+        let {propose, from_account, to_account, asset, asset_id, propose_account, feeAmount,
+            amount, error, to_name, from_name, memo, feeAsset, fee_asset_id, balanceError} = this.state;
         let from_my_account = AccountStore.isMyAccount(from_account) || from_name === this.props.passwordAccount;
 
         if(from_account && ! from_my_account && ! propose ) {
@@ -236,65 +331,27 @@ class Transfer extends React.Component {
             </span>;
         }
 
-        let fee = this._feeBTS = utils.estimateFee(propose ? "proposal_create" : "transfer", null, ChainStore.getObject("2.0.0"));
         let { asset_types, fee_asset_types } = this._getAvailableAssets();
         let balance = null;
 
         // Estimate fee
+        let fee = this.state.feeAmount.getAmount({real: true});
         if (from_account && from_account.get("balances") && !from_error) {
 
             let account_balances = from_account.get("balances").toJS();
-
-            // Finish fee estimation
-            let core = ChainStore.getObject("1.3.0");
-            if (feeAsset && feeAsset.get("id") !== "1.3.0" && core) {
-
-                let price = utils.convertPrice(core, feeAsset.getIn(["options", "core_exchange_rate"]).toJS(), null, feeAsset.get("id"));
-                fee = utils.convertValue(price, fee, core, feeAsset);
-
-                if (parseInt(fee, 10) !== fee) {
-                    fee += 1; // Add 1 to round up;
-                }
-            }
-            if (core) {
-                fee = utils.limitByPrecision(utils.get_asset_amount(fee, feeAsset || core), feeAsset ? feeAsset.get("precision") : core.get("precision"));
-            }
-
-            let current_asset_id = asset ? asset.get("id") : asset_types[0];
             if (asset_types.length === 1) asset = ChainStore.getAsset(asset_types[0]);
-            if (asset_types.length > 0) {                
+            if (asset_types.length > 0) {
+                let current_asset_id = asset ? asset.get("id") : asset_types[0];
                 let feeID = feeAsset ? feeAsset.get("id") : "1.3.0";
                 balance = (<span style={{borderBottom: "#A09F9F 1px dotted", cursor: "pointer"}} onClick={this._setTotal.bind(this, current_asset_id, account_balances[current_asset_id], fee, feeID)}><Translate component="span" content="transfer.available"/>: <BalanceComponent balance={account_balances[current_asset_id]}/></span>);
             } else {
                 balance = "No funds";
             }
-
-            fee_asset_types.push("1.3.0");
-            fee_asset_types = fee_asset_types.filter(e => {
-                let balanceObject = account_balances[e]?ChainStore.getObject(account_balances[e]):null;
-                let transferAsset = ChainStore.getObject(e);
-                let amount = 0;
-                if (balanceObject) {
-                    amount = utils.get_asset_amount(balanceObject.get("balance"), transferAsset);
-                } else {
-                    return e;
-                }
-                return amount - fee > fee;
-
-            });
-        } else {
-            let core = ChainStore.getObject("1.3.0");
-            fee_asset_types = ["1.3.0"];
-            if (core) {
-                fee = utils.limitByPrecision(utils.get_asset_amount(fee, feeAsset || core), feeAsset ? feeAsset.get("precision") : core.get("precision"));
-            }
-
         }
-
 
         let propose_incomplete = propose && ! propose_account;
         let submitButtonClass = "button float-right no-margin";
-        if(!from_account || !to_account || !amount || amount === "0"|| !asset || from_error || propose_incomplete)
+        if(!from_account || !to_account || !amount || amount === "0"|| !asset || from_error || propose_incomplete || balanceError)
             submitButtonClass += " disabled";
 
         let accountsList = Immutable.Set();
@@ -328,7 +385,7 @@ class Transfer extends React.Component {
                                 label="transfer.to"
                                 accountName={to_name}
                                 onChange={this.toChanged.bind(this)}
-                                onToAccountChanged={this.onToAccountChanged.bind(this)}
+                                onAccountChanged={this.onToAccountChanged.bind(this)}
                                 account={to_name}
                                 size={60}
                                 tabIndex={tabIndex++}
@@ -345,6 +402,7 @@ class Transfer extends React.Component {
                                 display_balance={balance}
                                 tabIndex={tabIndex++}
                             />
+                            {this.state.balanceError ? <p className="has-error no-margin" style={{paddingTop: 10}}><Translate content="transfer.errors.insufficient" /></p>:null}
                         </div>
                         {/*  M E M O  */}
                         <div className="content-block transfer-input">
@@ -368,9 +426,10 @@ class Transfer extends React.Component {
                                 disabled={true}
                                 amount={fee}
                                 onChange={this.onFeeChanged.bind(this)}
-                                asset={fee_asset_types.length && feeAsset ? feeAsset.get("id") : ( fee_asset_types.length === 1 ? fee_asset_types[0] : fee_asset_id ? fee_asset_id : fee_asset_types[0])}
+                                asset={fee_asset_types.length && feeAmount ? feeAmount.asset_id : ( fee_asset_types.length === 1 ? fee_asset_types[0] : fee_asset_id ? fee_asset_id : fee_asset_types[0])}
                                 assets={fee_asset_types}
                                 tabIndex={tabIndex++}
+                                error={this.state.hasPoolBalance === false ? "transfer.errors.insufficient" : null}
                             />
                             {propose ?
                                 <button className={submitButtonClass} type="submit" value="Submit" tabIndex={tabIndex++}>
