@@ -126,7 +126,7 @@ class ValueComponent extends React.Component {
 
         let eqValue = price ? utils.convertValue(price, amount, fromAsset, toAsset) : null;
 
-        if (!eqValue) {
+        if (!eqValue && eqValue !== 0) {
             return <div className="tooltip inline-block" data-place="bottom" data-tip={counterpart.translate("tooltip.no_price")} style={{fontSize: "0.9rem"}}><Translate content="account.no_price" /></div>;
         }
 
