@@ -1,6 +1,7 @@
 import {ChainStore} from "bitsharesjs/es";
 import utils from "./utils";
 import counterpart from "counterpart";
+import { estimateFee } from "./trxHelper";
 
 
 let scamAccountsPolo = [
@@ -81,7 +82,7 @@ export default class AccountUtils {
 
         let globalObject = ChainStore.getObject("2.0.0");
 
-        let fee = utils.estimateFee(operation, null, globalObject);
+        let fee = estimateFee(operation, null, globalObject);
 
         let accountBalances = account.get("balances");
         if (!accountBalances) {
