@@ -76,13 +76,14 @@ module.exports = function() {
 
         // WRAP INTO CSS FILE
         const extractCSS = new ExtractTextPlugin("app.css");
-        cssLoaders = extractCSS.extract({fallbackLoader: "style-loader",
-            loader: [{loader: "css-loader"}, {loader: "postcss-loader", options: {
+        cssLoaders = ExtractTextPlugin.extract({
+            fallback: "style-loader",
+            use: [{loader: "css-loader"}, {loader: "postcss-loader", options: {
                 plugins: [require("autoprefixer")]
             }}]}
         );
-        scssLoaders = extractCSS.extract({fallbackLoader: "style-loader",
-            loader: [{loader: "css-loader"}, {loader: "postcss-loader", options: {
+        scssLoaders = ExtractTextPlugin.extract({fallback: "style-loader",
+            use: [{loader: "css-loader"}, {loader: "postcss-loader", options: {
                 plugins: [require("autoprefixer")]
             }}, {loader: "sass-loader", options: {outputStyle: "expanded"}}]}
         );
@@ -98,6 +99,7 @@ module.exports = function() {
             minimize: true,
             debug: false
         }));
+        plugins.push(new webpack.optimize.ModuleConcatenationPlugin());
         if (!env.noUgly) {
 
             plugins.push(new webpack.optimize.UglifyJsPlugin({
@@ -164,7 +166,7 @@ module.exports = function() {
                 { test: /\.(coffee\.md|litcoffee)$/, loader: "coffee-loader?literate" },
                 {
                     test: /\.css$/,
-                    loader: cssLoaders
+                    use: cssLoaders
                 },
 
                 // var cssLoaders = "style-loader!css-loader!postcss-loader",
@@ -172,7 +174,7 @@ module.exports = function() {
 
                 {
                     test: /\.scss$/,
-                    loader: scssLoaders
+                    use: scssLoaders
                 },
                 {
                     test: /\.png$/,
